Fix CI Issue and Improvement of test
