--- bam/deployment/src/test/java/com/ibm/generativeai/bam/deployment/Util.java
@@ -1,9 +1,0 @@
-package com.ibm.generativeai.bam.deployment;
-
-public class Util {
-
-    public static String URL = "http://localhost:8089";
-    public static String URL_CHAT_API = "/v2/text/chat?version=%s";
-    public static int PORT = 8089;
-    public static String API_KEY = "my_super_token";
-}
